refactor: 기존 게시판 생성 API 리팩터링
