+++ prestudy/domain/post/adapter-out/src/main/kotlin/y/june/prestudy/post/persistence/entity/PostJpaEntity.kt
@@ -9,7 +9,7 @@ import java.time.LocalDateTime
 class PostJpaEntity(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long? = null,
+    val id: Long = 0,
     val author: String,
     val title: String,
     val password: String,
