+++ prestudy/domain/post/application/src/main/kotlin/y/june/prestudy/post/model/Post.kt
@@ -3,7 +3,7 @@ package y.june.prestudy.post.model
 import java.time.LocalDateTime
 
 class Post(
-    val id: Long? = null,
+    val id: Long = 0,
     val author: String,
     val title: String,
     val password: String,
